+++ packages/client/src/widgets/QuestionDetail/ui/QuestionInformation.tsx
