+++ packages/client/src/widgets/QuestionDetail/ui/QuestionInformationBottom.tsx
