+++ packages/client/src/widgets/QuestionDetail/ui/questionInformation.css.ts
